Rename footer to border_text and add text to it
