+++ zuckbot/gamestates/game_screen.py
@@ -2,7 +2,7 @@ import pygame
 
 from .objects.question_input import Question_Input
 from .objects.zuckbot import Zuckbot
-from .objects.footer import Footer
+from .objects.border_text import Border_Text
 
 
 class Game_Screen:
@@ -18,7 +18,7 @@ class Game_Screen:
 
         self.question_input = Question_Input(self)
         self.zuckbot = Zuckbot(self)
-        self.footer = Footer(self)
+        self.border_text = Border_Text(self)
 
         self.is_running = False
 
@@ -69,6 +69,6 @@ class Game_Screen:
 
             self.question_input.blitme()
             self.zuckbot.blitme()
-            self.footer.blitme()
+            self.border_text.blitme()
 
             pygame.display.flip()
